# docs/swagger.yaml
definitions:
  errcode.Error:
    properties:
      code:
        type: integer
      data: {}
      message:
        description: 用于存储自定义消息
        type: string
    type: object
...
  model.DeviceAuthReq:
    properties:
      device_name:
        description: 设备名称(可选)
        maxLength: 255
        type: string
      device_number:
        description: 设备唯一标识
        maxLength: 255
        type: string
      product_key:
        description: 产品密钥(可选，用于产品关联)
        maxLength: 255
        type: string
      template_secret:
        description: 模板密钥
        maxLength: 255
        type: string
    required:
    - device_number
    - template_secret
    type: object
  model.DeviceAuthRes:
    properties:
      device_id:
        description: 设备ID
        type: string
      voucher:
        description: 设备凭证
        type: string
    type: object
...
host: localhost:9999
info:
  contact: {}
  description: ThingsPanel API.
  title: ThingsPanel API
  version: "1.0"
paths:
...
  /api/v1/device/auth:
    post:
      consumes:
      - application/json
      description: 实现一型一密认证机制，设备通过此接口获取凭证
      parameters:
      - description: 认证请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DeviceAuthReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.DeviceAuthRes'
        "400":
          description: 错误响应
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 设备动态认证
      tags:
      - 设备认证
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"